FAIT-Calculators/Home-Depot-Scraper/checkout/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
stripe==5.4.0
python-dotenv==1.0.0
google-cloud-firestore==2.11.1
pydantic==1.10.7
jinja2==3.1.2
aiofiles==23.1.0
